+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/bookmark/BookmarkFragment.kt
@@ -3,16 +3,15 @@ package au.com.shiftyjelly.pocketcasts.player.view.bookmark
 import android.app.Activity
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.compose.foundation.background
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.fragment.compose.content
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.models.entity.Bookmark
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
@@ -27,26 +26,20 @@ class BookmarkFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View {
-        viewModel.load(BookmarkArguments.createFromArguments(arguments))
-        return ComposeView(requireContext()).apply {
-            setContent {
-                AppThemeWithBackground(Theme.ThemeType.DARK) {
-                    setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-
-                    val uiState: BookmarkViewModel.UiState by viewModel.uiState.collectAsState()
-                    BookmarkPage(
-                        isNewBookmark = uiState.isNewBookmark,
-                        title = uiState.title,
-                        tintColor = uiState.tintColor,
-                        backgroundColor = uiState.backgroundColor,
-                        onTitleChange = { viewModel.changeTitle(it) },
-                        onSave = ::saveBookmark,
-                        onClose = ::close,
-                        modifier = Modifier.background(uiState.backgroundColor),
-                    )
-                }
-            }
+    ) = content {
+        LaunchedEffect(Unit) { viewModel.load(BookmarkArguments.createFromArguments(arguments)) }
+        AppThemeWithBackground(Theme.ThemeType.DARK) {
+            val uiState: BookmarkViewModel.UiState by viewModel.uiState.collectAsState()
+            BookmarkPage(
+                isNewBookmark = uiState.isNewBookmark,
+                title = uiState.title,
+                tintColor = uiState.tintColor,
+                backgroundColor = uiState.backgroundColor,
+                onTitleChange = { viewModel.changeTitle(it) },
+                onSave = ::saveBookmark,
+                onClose = ::close,
+                modifier = Modifier.background(uiState.backgroundColor),
+            )
         }
     }