+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/entity/Podcast.kt
@@ -155,6 +155,12 @@ data class Podcast(
     val isAutoAddToUpNextOff: Boolean
         get() = autoAddToUpNext == AutoAddUpNext.OFF
 
+    val isAutoAddToUpNextPlayLast: Boolean
+        get() = autoAddToUpNext == AutoAddUpNext.PLAY_LAST
+
+    val isAutoAddToUpNextPlayNext: Boolean
+        get() = autoAddToUpNext == AutoAddUpNext.PLAY_NEXT
+
     val adapterId: Long
         get() = UUID.nameUUIDFromBytes(uuid.toByteArray()).mostSignificantBits