+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/CachedServerCallback.kt
@@ -1,8 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.servers
-
-interface CachedServerCallback<T> {
-
-    fun cachedDataFound(data: T)
-    fun networkDataFound(data: T)
-    fun notFound()
-}