+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/ServiceManager.kt
@@ -98,7 +98,7 @@ open class ServiceManager @Inject constructor(
     }
 
     fun exportFeedUrls(uuids: List<String>, callback: ServerCallback<Map<String, String>>): Call? {
-        val uuidsJoined = TextUtils.join(",", uuids)
+        val uuidsJoined = TextUtils.join(LIST_SEPERATOR, uuids)
         return postToMainServer(
             "/import/export_feed_urls",
             Parameters("uuids", uuidsJoined),
@@ -225,7 +225,6 @@ open class ServiceManager @Inject constructor(
         val formBody = parameters.toFormBody()
         val request = Request.Builder()
             .url(url)
-            .header("User-Agent", Settings.USER_AGENT_POCKETCASTS_SERVER)
             .post(formBody)
             .build()
 
@@ -368,7 +367,7 @@ open class ServiceManager @Inject constructor(
 
         operator fun get(key: String) = pairs.associate { (first, _) ->
             first to pairs.filter { it.first == first }.map { it.second }
-        }[key]?.joinToString(",")
+        }[key]?.joinToString(LIST_SEPERATOR)
 
         fun toFormBody(): FormBody {
             val builder = FormBody.Builder()